+++ CHANGELOG.markdown
@@ -14,6 +14,7 @@
 - Redesigned progress bar API
 - Redesigned color editor/picker API
 - Redesigned child window API (previously known as child frame)
+- Redesigned image / image button API
 - Updated layout API
 - Renderer errors implement std::error::Error
 - Glium renderer re-exports imgui and glium